[#10] /player Endpoint
- :globe_with_meridians: created controller
- :file_cabinet: database structure
